// frontend/bindings/github.com/telesma-app/kit/model/credentials/models.ts
export class CredentialRecord {
    "credentialIDHex": string;
    "credentialType"?: string;
    "credentialTransports"?: string[];
    "userIDHex"?: string;
    "userName"?: string;
    "displayName"?: string;
    "credProtect"?: number;
    "largeBlobKeyState"?: string;
    "thirdPartyPayment"?: boolean | null;

    /** Creates a new CredentialRecord instance. */
    constructor($$source: Partial<CredentialRecord> = {}) {
        if (!("credentialIDHex" in $$source)) {
            this["credentialIDHex"] = "";
        }

        Object.assign(this, $$source);
    }

    /**
     * Creates a new CredentialRecord instance from a string or object.
     */
    static createFrom($$source: any = {}): CredentialRecord {
        const $$createField2_0 = $$createType2;
        let $$parsedSource = typeof $$source === 'string' ? JSON.parse($$source) : $$source;
        if ("credentialTransports" in $$parsedSource) {
            $$parsedSource["credentialTransports"] = $$createField2_0($$parsedSource["credentialTransports"]);
        }
        return new CredentialRecord($$parsedSource as Partial<CredentialRecord>);
    }
}

export class CredentialTarget {
    "record": CredentialRecord;
    "rp": RelyingParty;
    "user": UserIdentity;

    /** Creates a new CredentialTarget instance. */
    constructor($$source: Partial<CredentialTarget> = {}) {
        if (!("record" in $$source)) {
            this["record"] = (new CredentialRecord());
        }
        if (!("rp" in $$source)) {
            this["rp"] = (new RelyingParty());
        }
        if (!("user" in $$source)) {
            this["user"] = (new UserIdentity());
        }

        Object.assign(this, $$source);
    }

    /**
     * Creates a new CredentialTarget instance from a string or object.
     */
    static createFrom($$source: any = {}): CredentialTarget {
        const $$createField0_0 = $$createType0;
        const $$createField1_0 = $$createType3;
        const $$createField2_0 = $$createType4;
        let $$parsedSource = typeof $$source === 'string' ? JSON.parse($$source) : $$source;
        if ("record" in $$parsedSource) {
            $$parsedSource["record"] = $$createField0_0($$parsedSource["record"]);
        }
        if ("rp" in $$parsedSource) {
            $$parsedSource["rp"] = $$createField1_0($$parsedSource["rp"]);
        }
        if ("user" in $$parsedSource) {
            $$parsedSource["user"] = $$createField2_0($$parsedSource["user"]);
        }
        return new CredentialTarget($$parsedSource as Partial<CredentialTarget>);
    }
}
// ...
export class RelyingParty {
    "id": string;
    "name"?: string;
    "idHashHex"?: string;

    /** Creates a new RelyingParty instance. */
    constructor($$source: Partial<RelyingParty> = {}) {
        if (!("id" in $$source)) {
            this["id"] = "";
        }

        Object.assign(this, $$source);
    }

    /**
     * Creates a new RelyingParty instance from a string or object.
     */
    static createFrom($$source: any = {}): RelyingParty {
        let $$parsedSource = typeof $$source === 'string' ? JSON.parse($$source) : $$source;
        return new RelyingParty($$parsedSource as Partial<RelyingParty>);
    }
}
// ...
export class UserIdentity {
    "userIDHex"?: string;
    "name"?: string;
    "displayName"?: string;

    /** Creates a new UserIdentity instance. */
    constructor($$source: Partial<UserIdentity> = {}) {

        Object.assign(this, $$source);
    }

    /**
     * Creates a new UserIdentity instance from a string or object.
     */
    static createFrom($$source: any = {}): UserIdentity {
        let $$parsedSource = typeof $$source === 'string' ? JSON.parse($$source) : $$source;
        return new UserIdentity($$parsedSource as Partial<UserIdentity>);
    }
}

// Private type creation functions
const $$createType0 = CredentialRecord.createFrom;
const $$createType1 = $Create.Array($$createType0);
const $$createType2 = $Create.Array($Create.Any);
const $$createType3 = RelyingParty.createFrom;
const $$createType4 = UserIdentity.createFrom;